--- __tests__/Antarctica.js
@@ -1,23 +1,0 @@
-// import 'react-native';
-// import React from 'react';
-// import jest from 'jest';
-// import Antarctica from '../components/Antarctica';
-
-// jest.unmock('./../../lib/components/webView');
-
-// Note: test renderer must be required after react-native.
-// import renderer from 'react-test-renderer';
-
-
-
-// jest.mock('./../../lib/components/webView', () => {
-//     const RN = require('react-native');
-//     return RN.Text;
-// });
-
-// test('renders correctly', () => {
-//     const tree = renderer.create(
-//         <Antarctica />
-//     ).toJSON();
-//     expect(tree).toMatchSnapshot();
-// });
