+++ __tests__/DinoQuiz.js
@@ -1,13 +1,13 @@
 import 'react-native';
 import React from 'react';
-import DinoQuiz from '../components/DinoQuiz';
+import AntarcticaQuiz from '../components/AntarcticaQuiz';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
 
 test('renders correctly', () => {
     const tree = renderer.create(
-        <DinoQuiz />
+        <AntarcticaQuiz />
     ).toJSON();
     expect(tree).toMatchSnapshot();
 });
