+++ __tests__/Splash.js
@@ -1,6 +1,6 @@
 import 'react-native';
 import React from 'react';
-import Splash from '../components/Splash';
+import Splash from '../src/components/Splash';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
