+++ src/components/Splash.js
