+++ src/components/take_quiz.png
