+++ src/pages/Antarctica.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { StyleSheet, Text, View, TouchableHighlight, WebView, Linking, Image } from 'react-native';
 
-class Space extends Component
+class Antarctica extends Component
 {
     constructor (props)
     {
@@ -22,26 +22,26 @@ class Space extends Component
     {
         return (
             <View style={styles.container}>
-                <Image style={styles.image} source={require('./space_btn.png')}/>
+                <Image style={{width: 360}} source={require('../components/ant1.png')}/>
 
                 <WebView
-                    source={{html: "<html><body><iframe width='350' height='250' src='https://www.youtube.com/embed/gmXW3AdI3XA' frameborder='0'></iframe></body></html>"}}
+                    source={{html: "<html><body><iframe width='350' height='250' src='https://www.youtube.com/embed/GBJDiFJZf-A' frameborder='0'></iframe></body></html>"}}
                     javaScriptEnabled={true}
                     domStorageEnabled={true}
                     style={styles.video}
                 />
 
-                <TouchableHighlight style={styles.button} onPress={() => Linking.openURL('https://www.youtube.com/embed/gmXW3AdI3XA')}>
-                    <Image style={styles.image} source={require('./open360.png')}/>
+                <TouchableHighlight style={styles.button} onPress={() => Linking.openURL('https://www.youtube.com/embed/GBJDiFJZf-A')}>
+                    <Image style={styles.image} source={require('../components/open360.png')}/>
                 </TouchableHighlight>
 
                 <View style={styles.containerStyle}>
-                    <TouchableHighlight style={styles.button} onPress={() => this.navigate('spaceQuizPage')}>
-                        <Image style={styles.image} source={require('./take_quiz.png')}/>
+                    <TouchableHighlight style={styles.button} onPress={() => this.navigate('antarcticaQuizPage')}>
+                        <Image style={styles.image} source={require('../components/take_quiz.png')}/>
                     </TouchableHighlight>
 
                     <TouchableHighlight style={styles.button} onPress={() => {this.props.navigator.pop()}}>
-                        <Image style={styles.image} source={require('./back.png')}/>
+                        <Image style={styles.image} source={require('../components/back.png')}/>
                     </TouchableHighlight>
                 </View>
             </View>
@@ -56,7 +56,7 @@ const styles = StyleSheet.create(
                 flex: 1,
                 justifyContent: 'flex-start',
                 alignItems: 'center',
-                backgroundColor: '#9b59b6',
+                backgroundColor: '#1abc9c',
             },
         button:
             {
@@ -86,7 +86,7 @@ const styles = StyleSheet.create(
         video:
             {
                 width: 365,
-                backgroundColor: '#9b59b6',
+                backgroundColor: '#1abc9c',
                 margin: 10,
                 shadowColor: '#000',
                 shadowOffset: {width: 1, height: 5},
@@ -96,13 +96,14 @@ const styles = StyleSheet.create(
             },
         image:
             {
-                borderColor: '#6c3384',
+                borderColor: '#107561',
                 borderWidth: 3,
                 shadowColor: '#000000',
                 shadowOffset: {width: 0, height: 3},
                 shadowOpacity: 1,
+                width: 350,
                 backgroundColor: '#000000',
             }
     });
 
-export default Space;
+export default Antarctica;
