+++ src/pages/quizzes/AntarcticaQuiz.js
@@ -2,16 +2,16 @@ import React, { Component } from 'react';
 import { StyleSheet, Text, View, TouchableHighlight, Image } from 'react-native';
 
 var questionNumber = 1;
-var oceanQuizScore = 0;
+var dinoQuizScore = 0;
 
-class OceanQuiz extends Component
+class AntarcticaQuiz extends Component
 {
     constructor()
     {
         super();
         this.state =
             {
-                questionText: 'Around 50% of the Earth\'s surface is covered by oceans',
+                questionText: 'Antarctica is considered a desert',
                 nextBtnTxt: 'Next Question'
             }
     }
@@ -20,21 +20,21 @@ class OceanQuiz extends Component
     {
         if(questionNumber == 1)
         {
-            this.setState({questionText: 'WRONG: Around 70% of the Earth\'s surface is covered by oceans'});
+            this.setState({questionText: 'CORRECT!'});
+            dinoQuizScore++;
         }
         if(questionNumber == 2)
         {
-            this.setState({questionText: 'CORRECT!'});
-            oceanQuizScore++;
+            this.setState({questionText: 'WRONG: Polar Bears live in the Arctic (the North Pole) so they never see Penguins'})
         }
         if(questionNumber == 3)
         {
-            this.setState({questionText: 'CORRECT!'});
-            oceanQuizScore++;
+            this.setState({questionText: 'WRONG: 1000s of people live and work at various research facilities in Antarctica'})
         }
         if(questionNumber == 4)
         {
-            this.setState({questionText: 'WRONG: There are 5 oceans covering the surface of our globe (Pacific, Atlantic, Indian, Arctic & Southern)'})
+            this.setState({questionText: 'CORRECT!'})
+            dinoQuizScore++;
         }
     }
 
@@ -42,47 +42,47 @@ class OceanQuiz extends Component
     {
         if(questionNumber == 1)
         {
-            this.setState({questionText: 'CORRECT!'});
-            oceanQuizScore++;
+            this.setState({questionText: 'WRONG: Because it experiences such little rain, Antarctica is considered a desert'})
             //return{background: #9adb8c #db8c8c
         }
         if(questionNumber == 2)
         {
-            this.setState({questionText: 'WRONG: The largest ocean on Earth is the Pacific Ocean'})
+            this.setState({questionText: 'CORRECT!'})
+            dinoQuizScore++;
         }
         if(questionNumber == 3)
         {
-            this.setState({questionText: 'WRONG: We have only explored about 5% of the Earth\'s oceans'})
+            this.setState({questionText: 'CORRECT!'})
+            dinoQuizScore++;
         }
         if(questionNumber == 4)
         {
-            this.setState({questionText: 'CORRECT!'});
-            oceanQuizScore++;
+            this.setState({questionText: 'WRONG: Antarctica is bigger than Europe'})
         }
     }
 
     updateQuestion = () =>
     {
         questionNumber++;
-        if(questionNumber == 2) {this.setState({questionText: 'The largest ocean on Earth is the Pacific Ocean'})}
-        if(questionNumber == 3) {this.setState({questionText: 'We have only explored about 5% of the Earth\'s oceans'})}
-        if(questionNumber == 4) {this.setState({questionText: 'There are 4 oceans covering the surface of our globe'})}
+        if(questionNumber == 2) {this.setState({questionText: 'Polar Bears and Penguins live together in Antarctica'})}
+        if(questionNumber == 3) {this.setState({questionText: 'No humans live in Antarctica'})}
+        if(questionNumber == 4) {this.setState({questionText: 'Antarctica is bigger than Europe'})}
         if(questionNumber == 5)
         {
-            this.setState({questionText: 'Your Score: '+(oceanQuizScore/4*100)+"%", nextBtnTxt: 'Retry?'});
+            this.setState({questionText: 'Your Score: '+(dinoQuizScore/4*100)+"%", nextBtnTxt: 'Retry?'});
         }
         if(questionNumber > 5)
         {
             questionNumber = 1;
-            oceanQuizScore = 0;
-            this.setState({questionText: 'Around 50% of the Earth\'s surface is covered by oceans', nextBtnTxt: 'Next Question'});
+            dinoQuizScore = 0;
+            this.setState({questionText: 'Antarctica is considered a desert.', nextBtnTxt: 'Next Question'});
         }
     }
 
     resetVariables = () =>
     {
         questionNumber = 1;
-        oceanQuizScore = 0;
+        dinoQuizScore = 0;
         this.props.navigator.pop();
     }
 
@@ -90,19 +90,19 @@ class OceanQuiz extends Component
     {
         return (
             <View style={styles.container}>
-                <Image style={{width: 360}} source={require('./true_false.png')}/>
+                <Image style={{width: 360}} source={require('../components/true_false.png')}/>
 
                 <View style={styles.question}>
-                    <Text style={styles.questionText}> {this.state.questionText} </Text>
+                    <Image style={styles.image} source={require('../components/true.png')}/>
                 </View>
 
                 <View style={styles.optionsStyle}>
                     <TouchableHighlight style={styles.optButton} onPress = {this.updateTrue}>
-                        <Image style={styles.image} source={require('./true.png')}/>
+                        <Image style={styles.image} source={require('../components/true.png')}/>
                     </TouchableHighlight>
 
                     <TouchableHighlight style={styles.optButton} onPress = {this.updateFalse}>
-                        <Image style={styles.image} source={require('./false.png')}/>
+                        <Image style={styles.image} source={require('../components/false.png')}/>
                     </TouchableHighlight>
                 </View>
 
@@ -111,7 +111,7 @@ class OceanQuiz extends Component
                 </TouchableHighlight>
 
                 <TouchableHighlight style={styles.button} onPress={this.resetVariables}>
-                    <Image style={styles.image} source={require('./back.png')}/>
+                    <Image style={styles.image} source={require('../components/back.png')}/>
                 </TouchableHighlight>
             </View>
         );
@@ -125,26 +125,10 @@ const styles = StyleSheet.create(
                 flex: 1,
                 justifyContent: 'flex-start',
                 alignItems: 'center',
-                backgroundColor: '#3498db',
+                backgroundColor: '#1abc9c',
             },
         button:
             {
-                alignItems: 'center',
-                height: 50,
-                width: 350,
-                shadowColor: '#000',
-                shadowOffset: {width: 0, height: 3},
-                shadowOpacity: 0.5,
-                elevation: 2,
-                marginBottom: 10,
-                marginRight: 10,
-                marginLeft: 10
-            },
-        nextButton:
-            {
-                backgroundColor: '#00ff00',
-                borderColor: '#1d5a82',
-                borderWidth: 3,
                 alignItems: 'center',
                 height: 50,
                 width: 350,
@@ -167,6 +151,22 @@ const styles = StyleSheet.create(
                 paddingBottom: 10,
                 position: 'relative',
                 marginTop: 10
+            },
+        nextButton:
+            {
+                backgroundColor: '#00ff00',
+                borderColor: '#107561',
+                borderWidth: 3,
+                alignItems: 'center',
+                height: 50,
+                width: 350,
+                shadowColor: '#000',
+                shadowOffset: {width: 0, height: 3},
+                shadowOpacity: 0.5,
+                elevation: 2,
+                marginBottom: 10,
+                marginRight: 10,
+                marginLeft: 10
             },
         textStyle:
             {
@@ -210,7 +210,7 @@ const styles = StyleSheet.create(
             },
         image:
             {
-                borderColor: '#1d5a82',
+                borderColor: '#107561',
                 borderWidth: 3,
                 shadowColor: '#000000',
                 shadowOffset: {width: 0, height: 3},
@@ -219,4 +219,5 @@ const styles = StyleSheet.create(
             }
     });
 
-export default OceanQuiz;
+export default AntarcticaQuiz;
+
