+++ src/pages/quizzes/OceanQuiz.js
@@ -2,16 +2,16 @@ import React, { Component } from 'react';
 import { StyleSheet, Text, View, TouchableHighlight, Image } from 'react-native';
 
 var questionNumber = 1;
-var dinoQuizScore = 0;
+var oceanQuizScore = 0;
 
-class HumanBodyQuiz extends Component
+class OceanQuiz extends Component
 {
     constructor()
     {
         super();
         this.state =
             {
-                questionText: 'The smallest bone in the human body is located in the ear.',
+                questionText: 'Around 50% of the Earth\'s surface is covered by oceans',
                 nextBtnTxt: 'Next Question'
             }
     }
@@ -20,21 +20,21 @@ class HumanBodyQuiz extends Component
     {
         if(questionNumber == 1)
         {
-            this.setState({questionText: 'CORRECT!'});
-            dinoQuizScore++;
+            this.setState({questionText: 'WRONG: Around 70% of the Earth\'s surface is covered by oceans'});
         }
         if(questionNumber == 2)
         {
-            this.setState({questionText: 'WRONG: Humans have 300 bones at birth but this decreases to 206 bones by adulthood after some bones fuse together'})
+            this.setState({questionText: 'CORRECT!'});
+            oceanQuizScore++;
         }
         if(questionNumber == 3)
         {
-            this.setState({questionText: 'WRONG: The tongue has 8 muscles, so is not the strongest muscle in the body (The strongest is in your jaw and lets you chew)'})
+            this.setState({questionText: 'CORRECT!'});
+            oceanQuizScore++;
         }
         if(questionNumber == 4)
         {
-            this.setState({questionText: 'CORRECT!'})
-            dinoQuizScore++;
+            this.setState({questionText: 'WRONG: There are 5 oceans covering the surface of our globe (Pacific, Atlantic, Indian, Arctic & Southern)'})
         }
     }
 
@@ -42,47 +42,47 @@ class HumanBodyQuiz extends Component
     {
         if(questionNumber == 1)
         {
-            this.setState({questionText: 'WRONG: The smallest bone in the human body is located in the ear'})
+            this.setState({questionText: 'CORRECT!'});
+            oceanQuizScore++;
             //return{background: #9adb8c #db8c8c
         }
         if(questionNumber == 2)
         {
-            this.setState({questionText: 'CORRECT!'})
-            dinoQuizScore++;
+            this.setState({questionText: 'WRONG: The largest ocean on Earth is the Pacific Ocean'})
         }
         if(questionNumber == 3)
         {
-            this.setState({questionText: 'CORRECT!'})
-            dinoQuizScore++;
+            this.setState({questionText: 'WRONG: We have only explored about 5% of the Earth\'s oceans'})
         }
         if(questionNumber == 4)
         {
-            this.setState({questionText: 'WRONG: As well as having unique fingerprints, humans also have unique tongue prints'})
+            this.setState({questionText: 'CORRECT!'});
+            oceanQuizScore++;
         }
     }
 
     updateQuestion = () =>
     {
         questionNumber++;
-        if(questionNumber == 2) {this.setState({questionText: 'The human skeleton is made up of 100 bones'})}
-        if(questionNumber == 3) {this.setState({questionText: 'The tongue is the strongest muscle in the body.'})}
-        if(questionNumber == 4) {this.setState({questionText: 'As well as having unique fingerprints, humans also have unique tongue prints'})}
+        if(questionNumber == 2) {this.setState({questionText: 'The largest ocean on Earth is the Pacific Ocean'})}
+        if(questionNumber == 3) {this.setState({questionText: 'We have only explored about 5% of the Earth\'s oceans'})}
+        if(questionNumber == 4) {this.setState({questionText: 'There are 4 oceans covering the surface of our globe'})}
         if(questionNumber == 5)
         {
-            this.setState({questionText: 'Your Score: '+(dinoQuizScore/4*100)+"%", nextBtnTxt: 'Retry?'});
+            this.setState({questionText: 'Your Score: '+(oceanQuizScore/4*100)+"%", nextBtnTxt: 'Retry?'});
         }
         if(questionNumber > 5)
         {
             questionNumber = 1;
-            dinoQuizScore = 0;
-            this.setState({questionText: 'The smallest bone in the human body is located in the ear', nextBtnTxt: 'Next Question'});
+            oceanQuizScore = 0;
+            this.setState({questionText: 'Around 50% of the Earth\'s surface is covered by oceans', nextBtnTxt: 'Next Question'});
         }
     }
 
     resetVariables = () =>
     {
         questionNumber = 1;
-        dinoQuizScore = 0;
+        oceanQuizScore = 0;
         this.props.navigator.pop();
     }
 
@@ -90,7 +90,7 @@ class HumanBodyQuiz extends Component
     {
         return (
             <View style={styles.container}>
-                <Image style={{width: 360}} source={require('./true_false.png')}/>
+                <Image style={{width: 360}} source={require('../components/true_false.png')}/>
 
                 <View style={styles.question}>
                     <Text style={styles.questionText}> {this.state.questionText} </Text>
@@ -98,11 +98,11 @@ class HumanBodyQuiz extends Component
 
                 <View style={styles.optionsStyle}>
                     <TouchableHighlight style={styles.optButton} onPress = {this.updateTrue}>
-                        <Image style={styles.image} source={require('./true.png')}/>
+                        <Image style={styles.image} source={require('../components/true.png')}/>
                     </TouchableHighlight>
 
                     <TouchableHighlight style={styles.optButton} onPress = {this.updateFalse}>
-                        <Image style={styles.image} source={require('./false.png')}/>
+                        <Image style={styles.image} source={require('../components/false.png')}/>
                     </TouchableHighlight>
                 </View>
 
@@ -111,7 +111,7 @@ class HumanBodyQuiz extends Component
                 </TouchableHighlight>
 
                 <TouchableHighlight style={styles.button} onPress={this.resetVariables}>
-                    <Image style={styles.image} source={require('./back.png')}/>
+                    <Image style={styles.image} source={require('../components/back.png')}/>
                 </TouchableHighlight>
             </View>
         );
@@ -125,10 +125,26 @@ const styles = StyleSheet.create(
                 flex: 1,
                 justifyContent: 'flex-start',
                 alignItems: 'center',
-                backgroundColor: '#f1c40f',
+                backgroundColor: '#3498db',
             },
         button:
             {
+                alignItems: 'center',
+                height: 50,
+                width: 350,
+                shadowColor: '#000',
+                shadowOffset: {width: 0, height: 3},
+                shadowOpacity: 0.5,
+                elevation: 2,
+                marginBottom: 10,
+                marginRight: 10,
+                marginLeft: 10
+            },
+        nextButton:
+            {
+                backgroundColor: '#00ff00',
+                borderColor: '#1d5a82',
+                borderWidth: 3,
                 alignItems: 'center',
                 height: 50,
                 width: 350,
@@ -151,22 +167,6 @@ const styles = StyleSheet.create(
                 paddingBottom: 10,
                 position: 'relative',
                 marginTop: 10
-            },
-        nextButton:
-            {
-                backgroundColor: '#00ff00',
-                borderColor: '#d1850e',
-                borderWidth: 3,
-                alignItems: 'center',
-                height: 50,
-                width: 350,
-                shadowColor: '#000',
-                shadowOffset: {width: 0, height: 3},
-                shadowOpacity: 0.5,
-                elevation: 2,
-                marginBottom: 10,
-                marginRight: 10,
-                marginLeft: 10
             },
         textStyle:
             {
@@ -210,7 +210,7 @@ const styles = StyleSheet.create(
             },
         image:
             {
-                borderColor: '#d1850e',
+                borderColor: '#1d5a82',
                 borderWidth: 3,
                 shadowColor: '#000000',
                 shadowOffset: {width: 0, height: 3},
@@ -219,4 +219,4 @@ const styles = StyleSheet.create(
             }
     });
 
-export default HumanBodyQuiz;
+export default OceanQuiz;
