+++ src/pages/quizzes/VolcanoQuiz.js
@@ -1,19 +1,18 @@
 import React, { Component } from 'react';
-import { StyleSheet, Text, View, Image, TouchableHighlight } from 'react-native';
+import { StyleSheet, Text, View, TouchableHighlight, Image } from 'react-native';
 
 var questionNumber = 1;
 var dinoQuizScore = 0;
 
-class DinoQuiz extends Component
+class VolcanoQuiz extends Component
 {
     constructor()
     {
         super();
         this.state =
             {
-                questionText: 'Dinosaurs ruled the earth for over 160 million years',
-                nextBtnTxt: 'Next Question',
-                bgColor: '#db8c8c'
+                questionText: 'Magma and lava are made of the same thing',
+                nextBtnTxt: 'Next Question'
             }
     }
 
@@ -26,11 +25,11 @@ class DinoQuiz extends Component
         }
         if(questionNumber == 2)
         {
-            this.setState({questionText: 'WRONG: Modern-day birds descend from a type of dinosaurs known as theropods'})
+            this.setState({questionText: 'WRONG: The largest volcano in the Solar System is Olympus Mons on Mars'})
         }
         if(questionNumber == 3)
         {
-            this.setState({questionText: 'WRONG: Dinosaurs lived on Earth until around 65 million years ago when a mass extinction occurred'})
+            this.setState({questionText: 'WRONG: Volcanoes are also found on the ocean floor and even under icecaps'})
         }
         if(questionNumber == 4)
         {
@@ -43,7 +42,7 @@ class DinoQuiz extends Component
     {
         if(questionNumber == 1)
         {
-            this.setState({questionText: 'WRONG: Dinosaurs ruled the earth for over 160 million years'})
+            this.setState({questionText: 'WRONG: Both lava and magma are made from molten rock. It is called magma when it is underground and lava above ground'})
             //return{background: #9adb8c #db8c8c
         }
         if(questionNumber == 2)
@@ -58,16 +57,16 @@ class DinoQuiz extends Component
         }
         if(questionNumber == 4)
         {
-            this.setState({questionText: 'WRONG: There are currently over 330 described dinosaur species and this number is growing'})
+            this.setState({questionText: 'WRONG: Pumice is a volcanic rock that can float in water'})
         }
     }
 
     updateQuestion = () =>
     {
         questionNumber++;
-        if(questionNumber == 2) {this.setState({questionText: 'Modern-day wolves descend from a type of dinosaurs known as theropods'})}
-        if(questionNumber == 3) {this.setState({questionText: 'Dinosaurs lived on Earth until around 1 million years ago when a mass extinction occurred'})}
-        if(questionNumber == 4) {this.setState({questionText: 'There are currently over 330 described dinosaur species and this number is growing'})}
+        if(questionNumber == 2) {this.setState({questionText: 'The largest volcano in the Solar System is found on Earth'})}
+        if(questionNumber == 3) {this.setState({questionText: 'Volcanoes are only found above ground'})}
+        if(questionNumber == 4) {this.setState({questionText: 'Pumice is a volcanic rock that can float in water'})}
         if(questionNumber == 5)
         {
             this.setState({questionText: 'Your Score: '+(dinoQuizScore/4*100)+"%", nextBtnTxt: 'Retry?'});
@@ -76,7 +75,7 @@ class DinoQuiz extends Component
         {
             questionNumber = 1;
             dinoQuizScore = 0;
-            this.setState({questionText: 'Dinosaurs ruled the earth for over 160 million years', nextBtnTxt: 'Next Question'});
+            this.setState({questionText: 'Magma and lava are made of the same thing', nextBtnTxt: 'Next Question'});
         }
     }
 
@@ -91,7 +90,7 @@ class DinoQuiz extends Component
     {
         return (
             <View style={styles.container}>
-                <Image style={{width: 360}} source={require('./true_false.png')}/>
+                <Image style={{width: 360}} source={require('../components/true_false.png')}/>
 
                 <View style={styles.question}>
                     <Text style={styles.questionText}> {this.state.questionText} </Text>
@@ -99,11 +98,11 @@ class DinoQuiz extends Component
 
                 <View style={styles.optionsStyle}>
                     <TouchableHighlight style={styles.optButton} onPress = {this.updateTrue}>
-                        <Image style={styles.image} source={require('./true.png')}/>
+                        <Image style={styles.image} source={require('../components/true.png')}/>
                     </TouchableHighlight>
 
                     <TouchableHighlight style={styles.optButton} onPress = {this.updateFalse}>
-                        <Image style={styles.image} source={require('./false.png')}/>
+                        <Image style={styles.image} source={require('../components/false.png')}/>
                     </TouchableHighlight>
                 </View>
 
@@ -112,7 +111,7 @@ class DinoQuiz extends Component
                 </TouchableHighlight>
 
                 <TouchableHighlight style={styles.button} onPress={this.resetVariables}>
-                    <Image style={styles.image} source={require('./back.png')}/>
+                    <Image style={styles.image} source={require('../components/back.png')}/>
                 </TouchableHighlight>
             </View>
         );
@@ -126,7 +125,7 @@ const styles = StyleSheet.create(
                 flex: 1,
                 justifyContent: 'flex-start',
                 alignItems: 'center',
-                backgroundColor: '#2ecc71',
+                backgroundColor: '#e74c3c',
             },
         button:
             {
@@ -142,21 +141,21 @@ const styles = StyleSheet.create(
                 marginLeft: 10
             },
         nextButton:
-        {
-            backgroundColor: '#00ff00',
-            borderColor: '#1c8247',
-            borderWidth: 3,
-            alignItems: 'center',
-            height: 50,
-            width: 350,
-            shadowColor: '#000',
-            shadowOffset: {width: 0, height: 3},
-            shadowOpacity: 0.5,
-            elevation: 2,
-            marginBottom: 10,
-            marginRight: 10,
-            marginLeft: 10
-        },
+            {
+                backgroundColor: '#00ff00',
+                borderColor: '#962a1f',
+                borderWidth: 3,
+                alignItems: 'center',
+                height: 50,
+                width: 350,
+                shadowColor: '#000',
+                shadowOffset: {width: 0, height: 3},
+                shadowOpacity: 0.5,
+                elevation: 2,
+                marginBottom: 10,
+                marginRight: 10,
+                marginLeft: 10
+            },
         optButton:
             {
                 alignItems: 'center',
@@ -211,7 +210,7 @@ const styles = StyleSheet.create(
             },
         image:
             {
-                borderColor: '#1c8247',
+                borderColor: '#962a1f',
                 borderWidth: 3,
                 shadowColor: '#000000',
                 shadowOffset: {width: 0, height: 3},
@@ -220,4 +219,4 @@ const styles = StyleSheet.create(
             }
     });
 
-export default DinoQuiz;
+export default VolcanoQuiz;
